refract: error bars
